pyproject.toml
[project]
name = "pro1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.1.14",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]
